refactor: 消除编译警告（指针与变量的类型）

+++ code/camera.h
@@ -18,8 +18,8 @@
 #define center_x_set 89
 ///////////////////////////////////////////////////////////////////////////////
 extern uint8 image_copy[MT9V03X_H][MT9V03X_W];
-extern int center_x_offset; // 在中心线左为负数
-extern int center_y_diff;   // 绝对值
+extern float center_x_offset; // 在中心线左为负数
+extern float center_y_diff;   // 绝对值
 extern IfxCpu_mutexLock camera_mutex;
 
 void draw_center_line(uint8 *image);
